requirements.txt
streamlit==1.34.0
pandas==2.2.2
matplotlib==3.8.4
llama-cpp-python==0.2.68